MicroServicio Historial sin terminar y con cambios en el microServicio de reserva

+++ ms-historial-petpals/src/main/java/com/historial_petpals/repository/HistorialRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface HistorialRepo extends JpaRepository<Historial, Long> {
-    List<Historial> findByIdDueno(Long idDueno);
+    List<Historial> findByIdDueno(Integer idDueno);
     List<Historial> findByIdCuidador(Long idCuidador);
     List<Historial> findByIdMascota(Long idMascota);
     List<Historial> findByIdServicio(Long idServicio);
